requirements.txt
absl-py==2.3.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alphashape==1.3.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
charset-normalizer==3.4.3
click==8.3.0
click-log==0.4.0
cmake==4.1.2
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.2.0
datasets==4.1.1
decorator==5.2.1
dill==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
grpcio==1.76.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
ImageIO==2.37.2
importlib_metadata==8.7.0
ipython==8.37.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lit==18.1.8
llvmlite==0.45.0
Markdown==3.9
matplotlib==3.10.6
matplotlib-inline==0.2.1
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.10.1
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.62.0
numpy==1.26.4
open3d==0.19.0
opencv-python==4.12.0.88
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
plotly==6.3.1
plyfile==1.1.3
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==4.21.12
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
retrying==1.4.2
rpds-py==0.28.0
rtree==1.4.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.0.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.11.1
triton==2.0.0
tzdata==2025.2
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.15
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0